+++ restapi/models/VariantModel.py
@@ -7,11 +7,11 @@ variant = Table('variants', metadata,
     Column('name', String(50), nullable=True),
     Column('option', String(50), nullable=True),
     Column('price', BigInteger, nullable=False),
-    Column('stock', Integer, nullable=False),
+    Column('stock', BigInteger, nullable=False),
     Column('code', String(100), nullable=True),
     Column('barcode', String(100), nullable=True),
     Column('image', String(100), nullable=True),
     Column('discount', Integer, server_default=text("0")),
     Column('discount_active', Boolean, server_default=expression.false()),
-    Column('product_id',Integer,ForeignKey('products.id',onupdate='cascade',ondelete='cascade'),nullable=False)
+    Column('product_id',BigInteger,ForeignKey('products.id',onupdate='cascade',ondelete='cascade'),nullable=False)
 )
